feat(llm): renamed variant1 and variant2 into variantA and variantB
