+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantA/Details.tsx
