+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantA/Main.tsx
