+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantB/Details.tsx
