+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantB/Main.tsx
