+++ libs/ledgerjs/packages/types-live/src/analyticsOptInPrompt.ts
@@ -1,4 +1,4 @@
 export enum LlmAnalyticsOptInPromptVariants {
-  variant1 = 1,
-  variant2 = 2,
+  variantA = "A",
+  variantB = "B",
 }
