+++ libs/live-config/src/featureFlags/defaultFeatures.ts
@@ -434,7 +434,7 @@ export const DEFAULT_FEATURES: Features = {
   llmAnalyticsOptInPrompt: {
     enabled: false,
     params: {
-      variant: LlmAnalyticsOptInPromptVariants.variant1,
+      variant: LlmAnalyticsOptInPromptVariants.variantA,
     },
   },
 };
